Address code review findings for fetch implementation
Move LocalFetcher to _test.go (test double out of production binary),
add injectable http.Client with 30s timeout to HTTPFetcher, use
package-level interface conformance checks, strengthen test assertions
with errors.As and content validation.
