+++ cmd/vendor-profiles/fetch.go
@@ -4,19 +4,27 @@ import (
 	"fmt"
 	"io"
 	"net/http"
-	"os"
-	"path/filepath"
+	"time"
 )
 
 type Fetcher interface {
 	Fetch(repo, ref, source string) ([]byte, error)
 }
 
-type HTTPFetcher struct{}
+var _ Fetcher = (*HTTPFetcher)(nil)
+
+type HTTPFetcher struct {
+	Client *http.Client
+}
 
 func (f *HTTPFetcher) Fetch(repo, ref, source string) ([]byte, error) {
+	client := f.Client
+	if client == nil {
+		client = &http.Client{Timeout: 30 * time.Second}
+	}
+
 	url := fmt.Sprintf("https://raw.githubusercontent.com/%s/%s/%s", repo, ref, source)
-	resp, err := http.Get(url)
+	resp, err := client.Get(url)
 	if err != nil {
 		return nil, fmt.Errorf("GET %s: %w", url, err)
 	}
@@ -27,17 +35,4 @@ func (f *HTTPFetcher) Fetch(repo, ref, source string) ([]byte, error) {
 	}
 
 	return io.ReadAll(resp.Body)
-}
-
-type LocalFetcher struct {
-	BaseDir string
-}
-
-func (f *LocalFetcher) Fetch(repo, ref, source string) ([]byte, error) {
-	p := filepath.Join(f.BaseDir, source)
-	data, err := os.ReadFile(p)
-	if err != nil {
-		return nil, fmt.Errorf("reading local file %s: %w", p, err)
-	}
-	return data, nil
 }
